dm-integrity: Prevent RMW for full metadata sector writes
If full metadata sectors are being written, don't read them first. This
prevents a read-modify-write cycle and increases performance on HDDs.

To do this, now calculate the checksums for all sectors in the bio in one
go in integrity_metadata and then pass the result to dm_integrity_rw_tag,
which now checks if full sectors are overwritten using the new
dm_bufio_is_valid_partial_write function.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>