dm-bufio: Make dm_bufio_new and dm_bufio_mark_partial_buffer_dirty work together
Previously this would lead to data corruption, because the code
assumed that after dm_bufio_new the whole buffer would be overwritten.

Fix this by tracking if the buffer contains only partial data and
clean it up if the full buffer is requested.

To cleanup a partial buffer, first initiate a read to a shadow buffer
and wait until there are no external holders of the buffer to prevent
concurent writes to it. Then fill the unwritten areas with the read
data.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>