dm-bufio: Allow setting and getting write alignment
The default write alignment may not be optimal for all cases. Also
dm-bufio users may want to read the alignment to advertise a correct
optimal io size. This will be important for performance improvements
in the next commits.

Add a new parameter to dm_bufio_client_create and a function
dm_bufio_get_alignment to allow doing this.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>