dm-bufio: make dm_bufio_new and dm_bufio_mark_partial_buffer_dirty work together
Previously this would lead to data corruption, because the code
assumed that after dm_bufio_new the whole buffer would be overwritten.

Fix this by tracking if the buffer contains only partial data and
clean it up if the full buffer is requested. To allow non-sequential
writes to such a partial buffer, replace dirty_{start|end} and
write_{start|end} with bitmaps.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>