dm-bufio: Allow to check if exactly the specified part will be written
dm-bufio will align writes of partial dirty buffers and thus might
write more than what is asked. This may be undesirable if the buffer
that is being marked partial dirty was aquired with dm_bufio_new,
because then garbage may be written if the caller only wrote data
to the specified part.

To solve this, introduce a new function dm_bufio_is_valid_partial_write
which checks if exactly the specified part will be written.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>