dm-integrity: Flush data device before locking in bitmap_flush_work
bitmap_flush_work blocks writes to the whole device while it flushes
data and metadata and clears the bitmap. To reduce the time while
the lock is held, it flushes metadata before taking the lock.

For the same reason, also issue flush to the data device before
taking the lock.

Signed-off-by: Lukas Straub <lukasstraub2@web.de>